# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "summary": "Создать товар",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Товар успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "summary": "Получить список товаров",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Смещение для пагинации",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Поисковый запрос",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Поле для сортировки",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Направление сортировки: ASC или DESC",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список товаров получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProductResponse"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products/low-stock": {
      "get": {
        "operationId": "ProductsController_getLowStock",
        "summary": "Get products with low stock",
        "parameters": [
          {
            "name": "threshold",
            "required": false,
            "in": "query",
            "description": "Stock threshold (default 10)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max results (default 50, max 100)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Low stock products retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "summary": "Получить товар по ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Товар не найден"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "summary": "Обновить товар",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Товар успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Товар не найден"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "summary": "Удалить товар",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар успешно удален"
          },
          "400": {
            "description": "Невозможно удалить товар"
          },
          "404": {
            "description": "Товар не найден"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products/bulk": {
      "delete": {
        "operationId": "ProductsController_removeBulk",
        "summary": "Bulk delete products",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Products deleted"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ReceiptMaster API",
    "description": "Multi-tenant SaaS API for managing products, recipients, orders, receipts, and analytics.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название товара",
            "example": "iPhone 15"
          },
          "purchase_price_cents": {
            "type": "number",
            "description": "Цена покупки в копейках",
            "example": 80000
          },
          "sale_price_cents": {
            "type": "number",
            "description": "Цена продажи в копейках",
            "example": 99900
          },
          "quantity": {
            "type": "number",
            "description": "Количество товара на складе",
            "example": 50
          },
          "currency": {
            "type": "string",
            "description": "Валюта",
            "enum": [
              "UAH"
            ],
            "example": "UAH"
          }
        },
        "required": [
          "name",
          "purchase_price_cents",
          "sale_price_cents",
          "quantity",
          "currency"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "purchase_price_cents": {
            "type": "number"
          },
          "sale_price_cents": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "enum": [
              "UAH"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "purchase_price_cents",
          "sale_price_cents",
          "quantity",
          "currency",
          "created_at",
          "updated_at"
        ]
      },
      "PaginatedProductResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "data"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название товара",
            "example": "iPhone 15"
          },
          "purchase_price_cents": {
            "type": "number",
            "description": "Цена покупки в копейках",
            "example": 80000
          },
          "sale_price_cents": {
            "type": "number",
            "description": "Цена продажи в копейках",
            "example": 99900
          },
          "quantity": {
            "type": "number",
            "description": "Количество товара на складе",
            "example": 50
          },
          "currency": {
            "type": "string",
            "description": "Валюта",
            "enum": [
              "UAH"
            ],
            "example": "UAH"
          }
        }
      },
...
    }
  }
}