# backend/src/modules/products/entities/product.entity.ts
import {
  Entity,
  PrimaryGeneratedColumn,
  Column,
  CreateDateColumn,
  UpdateDateColumn,
  OneToMany,
  ManyToOne,
  JoinColumn,
  Index,
} from "typeorm";
import { ApiHideProperty, ApiProperty } from "@nestjs/swagger";
import { OrderItem } from "../../orders/entities/order-item.entity";
import { User } from "../../users/entities/user.entity";

export enum Currency {
  UAH = "UAH",
}

@Entity("products")
@Index(["user_id", "name"])
export class Product {
  @ApiProperty()
  @PrimaryGeneratedColumn("uuid")
  id: string;

  @ApiProperty()
  @Column({ type: "varchar", length: 255 })
  name: string;

  @ApiProperty()
  @Column({ type: "integer" })
  purchase_price_cents: number;

  @ApiProperty()
  @Column({ type: "integer" })
  sale_price_cents: number;

  @ApiProperty()
  @Column({ type: "integer", default: 0 })
  quantity: number;

  @ApiProperty({ enum: Currency })
  @Column({ type: "enum", enum: Currency, default: Currency.UAH })
  currency: Currency;

  @ApiHideProperty()
  @Column({ type: "uuid" })
  user_id: string;

  @ApiHideProperty()
  @ManyToOne(() => User, (user) => user.products)
  @JoinColumn({ name: "user_id" })
  user: User;

  @ApiProperty()
  @CreateDateColumn()
  created_at: Date;

  @ApiProperty()
  @UpdateDateColumn()
  updated_at: Date;

  @ApiHideProperty()
  @OneToMany(() => OrderItem, (orderItem) => orderItem.product)
  order_items: OrderItem[];
}